refactor: remove redundancies and centralize event management at `Skills` page
