+++ src/components/SkillsHeader/ISkillsHeader.ts
@@ -1,0 +1,5 @@
+import {ComponentProps, MouseEventHandler} from 'react'
+
+export interface ISkillsHeader extends ComponentProps<'header'> {
+    onHandleClick: MouseEventHandler<HTMLAnchorElement>
+}
