+++ src/components/SkillsList/ISkillsList.ts
@@ -1,0 +1,6 @@
+import {ComponentProps} from 'react'
+import {Skill} from '../../models'
+
+export interface ISkillsList extends ComponentProps<'ul'> {
+    skills: Skill[]
+}
