+++ src/components/SkillsList/index.ts
@@ -1,0 +1,1 @@
+export * from './ISkillsList'
